refactor readdir structs, remove duplicated header code

+++ win32/readdir.c
@@ -27,8 +27,8 @@ extern "C" {
 /* typedef DIR - not the same as Unix */
 struct DIR_W32 {
 	HANDLE handle;				/* _findfirst/_findnext handle */
-	int offset;					/* offset into directory */
-	short finished;				/* 1 if there are not more files */
+	uint16_t offset;					/* offset into directory */
+	uint8_t finished;				/* 1 if there are not more files */
 	WIN32_FIND_DATAW fileinfo;  /* from _findfirst/_findnext */
 	wchar_t *dirw;		/* the dir we are reading */
 	struct dirent dent;			/* the dirent to return */
